fix(): seperates login button and modal. Also closes model when esc is pressed and adds cursor styling for header buttons
