+++ src/components/ui/header.tsx
@@ -15,11 +15,11 @@ const Header: React.FC = () => {
 
 			{/* Buttons on the right */}
 			<div className="flex space-x-6 mr-6">
-				<button className="text-white flex items-center transition-transform duration-200 hover:scale-115">
+				<button className="text-white flex items-center transition-transform duration-200 hover:scale-115 cursor-pointer">
 					<FontAwesomeIcon icon={faEdit} className="mr-2" />
 					Edit Data
 				</button>
-				<button className="text-white flex items-center transition-transform duration-200 hover:scale-115">
+				<button className="text-white flex items-center transition-transform duration-200 hover:scale-115 cursor-pointer">
 					<FontAwesomeIcon icon={faUserShield} className="mr-2" />
 					Admin
 				</button>
